refactor(classes): Split odds and ends left in core file to their proper homes
Makes way for cleanup in SILE global needed for sane handling of new
require CLI flag.
